fix: rename hero.section.tsx file to hero-section.tsx
